GPU: Refactor some code, and keep only CPU-kernel realted code in GPUReconstructionCPU
