+++ GPU/GPUTracking/Base/GPUReconstructionKernels.h
@@ -24,20 +24,6 @@ namespace gpu
 
 namespace gpu_reconstruction_kernels
 {
-struct deviceEvent {
-  constexpr deviceEvent() = default;
-  constexpr deviceEvent(std::nullptr_t p) : v(nullptr){};
-  template <class T>
-  void set(T val) { v = reinterpret_cast<void*&>(val); }
-  template <class T>
-  T& get() { return reinterpret_cast<T&>(v); }
-  template <class T>
-  T* getEventList() { return reinterpret_cast<T*>(this); }
-  bool isSet() const { return v; }
-
- private:
-  void* v = nullptr; // We use only pointers anyway, and since cl_event and cudaEvent_t and hipEvent_t are actually pointers, we can cast them to deviceEvent (void*) this way.
-};
 
 template <class T, int32_t I = 0>
 struct classArgument {
@@ -95,6 +81,7 @@ struct krnlSetupArgs : public gpu_reconstruction_kernels::classArgument<T, I> {
   const krnlSetupTime s;
   std::tuple<typename std::conditional<(sizeof(Args) > sizeof(void*)), const Args&, const Args>::type...> v;
 };
+
 } // namespace gpu_reconstruction_kernels
 
 template <class T>
